Issue with requiring Terradactyl
Hey Nelson! 
Its Andrew. I tried to help Janyne analyze some data and ran into some walls with Devtools. Replicated errors on both of our computers. Traced it back to your addition of requiring terradactyl in the description file. Something was going awry there... this is not a solution, merely a fix as I am unsure of what is going wrong exactly.